fixup! Expose the record writer for validating arbitrary files
Move RecordWriterTestHarness into memray._memray

This is where we've put all of the test harnesses that depend upon our
C++ implementation, to avoid needing to build all of our C++ code
multiple times, and to avoid any risk of UB caused by multiple copies of
our C++ implementation being loaded in the same Python process (via
`memray._memray` and `memray._memray_test_utils`)
