+++ src/memray/_test.py
@@ -5,7 +5,6 @@ from ._test_utils import MmapAllocator
 from ._test_utils import PrimeCaches
 from ._test_utils import PymallocDomain
 from ._test_utils import PymallocMemoryAllocator
-from ._test_utils import RecordWriterTestHarness
 from ._test_utils import _cython_allocate_in_two_places
 from ._test_utils import _cython_nested_allocation
 from ._test_utils import allocate_cpp_vector
@@ -65,5 +64,4 @@ __all__ = [
     "fill_cpp_vector",
     "exit",
     "PrimeCaches",
-    "RecordWriterTestHarness",
 ]
